requirements.txt
box2d-py
graphviz
gym[atari]
nevergrad
numpy
tensorflow-gpu
tensorflow-probability
tensorboardX
torch
ray[tune]
